+++ extras/src/com/mopub/mobileads/TapjoyRewardedVideo.java
@@ -21,16 +21,23 @@ import com.tapjoy.TapjoyLog;
 import java.util.Hashtable;
 import java.util.Map;
 
-// Tested with Tapjoy SDK 11.5.1
+// Tested with Tapjoy SDK 11.8.2
 public class TapjoyRewardedVideo extends CustomEventRewardedVideo {
     private static final String TAG = TapjoyRewardedVideo.class.getSimpleName();
     private static final String TJC_MOPUB_NETWORK_CONSTANT = "mopub";
-    private static final String TJC_MOPUB_ADAPTER_VERSION_NUMBER = "4.0.0";
+    private static final String TJC_MOPUB_ADAPTER_VERSION_NUMBER = "4.1.0";
     private static final String TAPJOY_AD_NETWORK_CONSTANT = "tapjoy_id";
 
+    // Configuration keys
+    public static final String SDK_KEY = "sdkKey";
+    public static final String DEBUG_ENABLED = "debugEnabled";
+    public static final String PLACEMENT_NAME = "name";
+
     private String sdkKey;
+    private String placementName;
     private Hashtable<String, Object> connectFlags;
     private TJPlacement tjPlacement;
+    private boolean isAutoConnect = false;
     private static TapjoyRewardedVideoListener sTapjoyListener = new TapjoyRewardedVideoListener();
 
     static {
@@ -62,25 +69,33 @@ public class TapjoyRewardedVideo extends CustomEventRewardedVideo {
             @NonNull Map<String, String> serverExtras)
             throws Exception {
 
+        placementName = serverExtras.get(PLACEMENT_NAME);
+        if (TextUtils.isEmpty(placementName)) {
+            MoPubLog.d("Tapjoy rewarded video loaded with empty 'name' field. Request will fail.");
+        }
+
         if (!Tapjoy.isConnected()) {
             if (checkAndInitMediationSettings()) {
-                MoPubLog.d("Request to connect to Tapjoy");
+                MoPubLog.d("Connecting to Tapjoy via MoPub mediation settings...");
+                connectToTapjoy(launcherActivity);
 
-                Tapjoy.connect(launcherActivity, sdkKey, connectFlags, new TJConnectListener() {
-                    @Override
-                    public void onConnectSuccess() {
-                        MoPubLog.d("Tapjoy connected successfully");
-                    }
-
-                    @Override
-                    public void onConnectFailure() {
-                        MoPubLog.e("Tapjoy connect failed");
-                    }
-                });
-
+                isAutoConnect = true;
                 return true;
             } else {
-                MoPubLog.d("Cannot connect to Tapjoy -- missing 'sdkkey' declaration via TapjoyMediationSettings");
+                boolean enableDebug = Boolean.valueOf(serverExtras.get(DEBUG_ENABLED));
+                Tapjoy.setDebugEnabled(enableDebug);
+
+                sdkKey = serverExtras.get(SDK_KEY);
+                if (!TextUtils.isEmpty(sdkKey)) {
+                    MoPubLog.d("Connecting to Tapjoy via MoPub dashboard settings...");
+                    connectToTapjoy(launcherActivity);
+
+                    isAutoConnect = true;
+                    return true;
+                } else {
+                    MoPubLog.d("Tapjoy rewarded video is initialized with empty 'sdkKey'. You must call Tapjoy.connect()");
+                    isAutoConnect = false;
+                }
             }
         }
 
@@ -93,15 +108,40 @@ public class TapjoyRewardedVideo extends CustomEventRewardedVideo {
             @NonNull Map<String, String> serverExtras)
             throws Exception {
         MoPubLog.d("Requesting Tapjoy rewarded video");
+        createPlacement(activity);
+    }
 
-        String name = serverExtras.get("name");
-        if (TextUtils.isEmpty(name)) {
-            MoPubLog.d("Tapjoy interstitial loaded with empty 'name' field. Request will fail.");
+    private void connectToTapjoy(final Activity launcherActivity) {
+        Tapjoy.connect(launcherActivity, sdkKey, connectFlags, new TJConnectListener() {
+            @Override
+            public void onConnectSuccess() {
+                MoPubLog.d("Tapjoy connected successfully");
+                createPlacement(launcherActivity);
+            }
+
+            @Override
+            public void onConnectFailure() {
+                MoPubLog.d("Tapjoy connect failed");
+            }
+        });
+    }
+
+    private void createPlacement(Activity activity) {
+        if (!TextUtils.isEmpty(placementName)) {
+            if (isAutoConnect && !Tapjoy.isConnected()) {
+                // If adapter is making the Tapjoy.connect() call on behalf of the pub, wait for it to
+                // succeed before making a placement request.
+                MoPubLog.d("Tapjoy is still connecting. Please wait for this to finish before making a placement request");
+                return;
+            }
+
+            tjPlacement = new TJPlacement(activity, placementName, sTapjoyListener);
+            tjPlacement.setMediationName(TJC_MOPUB_NETWORK_CONSTANT);
+            tjPlacement.setAdapterVersion(TJC_MOPUB_ADAPTER_VERSION_NUMBER);
+            tjPlacement.requestContent();
+        } else {
+            MoPubLog.d("Tapjoy placementName is empty. Unable to create TJPlacement.");
         }
-        tjPlacement = new TJPlacement(activity, name, sTapjoyListener);
-        tjPlacement.setMediationName(TJC_MOPUB_NETWORK_CONSTANT);
-        tjPlacement.setAdapterVersion(TJC_MOPUB_ADAPTER_VERSION_NUMBER);
-        tjPlacement.requestContent();
     }
 
     @Override
@@ -117,19 +157,19 @@ public class TapjoyRewardedVideo extends CustomEventRewardedVideo {
         } else {
             MoPubLog.d("Failed to show Tapjoy rewarded video.");
         }
-
     }
 
     private boolean checkAndInitMediationSettings() {
-        MoPubLog.d("Initializing Tapjoy mediation settings");
-
         final TapjoyMediationSettings globalMediationSettings =
                 MoPubRewardedVideoManager.getGlobalMediationSettings(TapjoyMediationSettings.class);
 
         if (globalMediationSettings != null) {
+            MoPubLog.d("Initializing Tapjoy mediation settings");
+
             if (!TextUtils.isEmpty(globalMediationSettings.getSdkKey())) {
                 sdkKey = globalMediationSettings.getSdkKey();
             } else {
+                MoPubLog.d("Cannot initialize Tapjoy -- 'sdkkey' is empty");
                 return false;
             }
 
@@ -146,7 +186,7 @@ public class TapjoyRewardedVideo extends CustomEventRewardedVideo {
     private static class TapjoyRewardedVideoListener implements TJPlacementListener, CustomEventRewardedVideoListener, TJVideoListener {
         @Override
         public void onRequestSuccess(TJPlacement placement) {
-            if (!placement.isContentAvailable()){
+            if (!placement.isContentAvailable()) {
                 MoPubLog.d("No Tapjoy rewarded videos available");
                 MoPubRewardedVideoManager.onRewardedVideoLoadFailure(TapjoyRewardedVideo.class, TAPJOY_AD_NETWORK_CONSTANT, MoPubErrorCode.NETWORK_NO_FILL);
             }