+++ extras/src/com/mopub/nativeads/MillennialNative.java
@@ -2,6 +2,7 @@ package com.mopub.nativeads;
 
 import android.app.Activity;
 import android.content.Context;
+import android.os.Build;
 import android.os.Handler;
 import android.os.Looper;
 import android.util.Log;
@@ -11,6 +12,7 @@ import com.millennialmedia.AppInfo;
 import com.millennialmedia.MMException;
 import com.millennialmedia.MMSDK;
 import com.millennialmedia.NativeAd;
+import com.millennialmedia.internal.ActivityListenerManager;
 
 import java.util.ArrayList;
 import java.util.List;
@@ -21,42 +23,35 @@ import static com.mopub.nativeads.NativeImageHelper.preCacheImages;
 public class MillennialNative extends CustomEventNative {
     public static final String DCN_KEY = "dcn";
     public static final String APID_KEY = "adUnitID";
-    private final static String LOGCAT_TAG = "MoPub->MM-Native";
+    private final static String TAG = MillennialNative.class.getSimpleName();
     private static final Handler UI_THREAD_HANDLER = new Handler(Looper.getMainLooper());
 
     @Override
     protected void loadNativeAd(final Context context,
-            final CustomEventNativeListener listener,
-            final Map<String, Object> localExtras,
-            final Map<String, String> serverExtras) {
-
+            final CustomEventNativeListener customEventNativeListener,
+            Map<String, Object> localExtras,
+            Map<String, String> serverExtras) {
         String placementId;
         String siteId;
-
-        if ( !MMSDK.isInitialized() ) {
-            try {
-                MMSDK.initialize((Activity) context);
-            } catch ( Exception e ) {
-                Log.e(LOGCAT_TAG, "Unable to initialize the Millennial SDK-- " + e.getMessage());
-                e.printStackTrace();
-                UI_THREAD_HANDLER.post(new Runnable() {
-                    @Override
-                    public void run() {
-                        listener.onNativeAdFailed(NativeErrorCode.NATIVE_ADAPTER_CONFIGURATION_ERROR);
-                    }
-                });
-                return;
-            }
+        if (!initializeSDK(context)) {
+            Log.e(TAG, "Unable to initialize MMSDK");
+            UI_THREAD_HANDLER.post(new Runnable() {
+                @Override
+                public void run() {
+                    customEventNativeListener.onNativeAdFailed(NativeErrorCode.NATIVE_ADAPTER_CONFIGURATION_ERROR);
+                }
+            });
+            return;
         }
 
-        if ( extrasAreValid( serverExtras )) {
+        if (extrasAreValid(serverExtras)) {
             placementId = serverExtras.get(APID_KEY);
             siteId = serverExtras.get(DCN_KEY);
         } else {
             UI_THREAD_HANDLER.post(new Runnable() {
                 @Override
                 public void run() {
-                    listener.onNativeAdFailed(NativeErrorCode.NATIVE_ADAPTER_CONFIGURATION_ERROR);
+                    customEventNativeListener.onNativeAdFailed(NativeErrorCode.NATIVE_ADAPTER_CONFIGURATION_ERROR);
                 }
             });
             return;
@@ -64,19 +59,23 @@ public class MillennialNative extends CustomEventNative {
 
         try {
             AppInfo ai = new AppInfo().setMediator("mopubsdk");
-            if ( siteId != null && siteId.length() > 0 ) {
+            if (siteId != null && siteId.length() > 0) {
                 ai = ai.setSiteId(siteId);
             } else {
                 ai = ai.setSiteId(null);
             }
 
-            MMSDK.setAppInfo(ai);
-        } catch ( IllegalStateException e ) {
-            Log.w(LOGCAT_TAG, "Caught exception: " + e.getMessage());
+            try {
+                MMSDK.setAppInfo(ai);
+            } catch (MMException e) {
+                Log.e(TAG, "MM SDK is not initialized", e);
+            }
+        } catch (IllegalStateException e) {
+            Log.w(TAG, "App info error", e);
             UI_THREAD_HANDLER.post(new Runnable() {
                 @Override
                 public void run() {
-                    listener.onNativeAdFailed(NativeErrorCode.NATIVE_ADAPTER_CONFIGURATION_ERROR);
+                    customEventNativeListener.onNativeAdFailed(NativeErrorCode.NATIVE_ADAPTER_CONFIGURATION_ERROR);
                 }
             });
             return;
@@ -89,13 +88,13 @@ public class MillennialNative extends CustomEventNative {
                             nativeAd,
                             new ImpressionTracker(context),
                             new NativeClickHandler(context),
-                            listener);
+                            customEventNativeListener);
             millennialStaticNativeAd.loadAd();
-        } catch ( MMException e ) {
+        } catch (MMException e) {
             UI_THREAD_HANDLER.post(new Runnable() {
                 @Override
                 public void run() {
-                    listener.onNativeAdFailed(NativeErrorCode.NATIVE_ADAPTER_CONFIGURATION_ERROR);
+                    customEventNativeListener.onNativeAdFailed(NativeErrorCode.NATIVE_ADAPTER_CONFIGURATION_ERROR);
                 }
             });
         }
@@ -104,7 +103,7 @@ public class MillennialNative extends CustomEventNative {
     private boolean extrasAreValid(final Map<String, String> serverExtras) {
         String placementId = serverExtras.get(APID_KEY);
         return (serverExtras.containsKey(APID_KEY) &&
-                placementId != null && placementId.length() > 0 );
+                placementId != null && placementId.length() > 0);
     }
 
 
@@ -132,12 +131,11 @@ public class MillennialNative extends CustomEventNative {
         }
 
         void loadAd() {
-            Log.i(LOGCAT_TAG, "Loading native ad...");
+            Log.d(TAG, "Millennial native ad loading.");
             try {
                 mNativeAd.load(mContext, null);
             } catch (MMException e) {
-                Log.w(MillennialNative.LOGCAT_TAG, "Caught configuration error Exception.");
-                e.printStackTrace();
+                Log.w(TAG, "Configuration error", e);
                 UI_THREAD_HANDLER.post(new Runnable() {
                     @Override
                     public void run() {
@@ -177,9 +175,9 @@ public class MillennialNative extends CustomEventNative {
             notifyAdImpressed();
             try {
                 mNativeAd.fireImpression();
-                Log.i(LOGCAT_TAG, "Millennial native impression recorded.");
-            } catch ( MMException m ) {
-                Log.e(LOGCAT_TAG, "Millennial native impression NOT tracked: " + m.getMessage() );
+                Log.d(TAG, "Millennial native ad impression recorded.");
+            } catch (MMException m) {
+                Log.e(TAG, "Error tracking Millennial native ad impression", m);
             }
         }
 
@@ -187,8 +185,8 @@ public class MillennialNative extends CustomEventNative {
         public void handleClick(final View view) {
             notifyAdClicked();
             mNativeClickHandler.openClickDestinationUrl(getClickDestinationUrl(), view);
-            mNativeAd.fireClicked();
-            Log.i(LOGCAT_TAG, "Millennial native ad clicked!");
+            mNativeAd.fireCallToActionClicked();
+            Log.d(TAG, "Millennial native ad clicked.");
         }
 
         // MM'S Native mListener
@@ -207,8 +205,8 @@ public class MillennialNative extends CustomEventNative {
                 UI_THREAD_HANDLER.post(new Runnable() {
                     @Override
                     public void run() {
-                        Log.d(LOGCAT_TAG,
-                                "Millennial native encountered null destination url. Failing over.");
+                        Log.d(TAG,
+                                "Millennial native ad encountered null destination url. Failing over.");
                         mListener.onNativeAdFailed(
                                 NativeErrorCode.NATIVE_ADAPTER_CONFIGURATION_ERROR);
                     }
@@ -220,9 +218,13 @@ public class MillennialNative extends CustomEventNative {
             setIconImageUrl(iconImageUrl);
             setMainImageUrl(mainImageUrl);
 
-            final List<String> urls = new ArrayList<String>();
-            if ( iconImageUrl != null ) { urls.add(iconImageUrl); }
-            if ( mainImageUrl != null ) { urls.add(mainImageUrl); }
+            final List<String> urls = new ArrayList<>();
+            if (iconImageUrl != null) {
+                urls.add(iconImageUrl);
+            }
+            if (mainImageUrl != null) {
+                urls.add(mainImageUrl);
+            }
 
             UI_THREAD_HANDLER.post(new Runnable() {
                 @Override
@@ -232,7 +234,7 @@ public class MillennialNative extends CustomEventNative {
                         @Override
                         public void onImagesCached() {
                             mListener.onNativeAdLoaded(mMillennialStaticNativeAd);
-                            Log.i(LOGCAT_TAG, "Millennial native ad loaded");
+                            Log.d(TAG, "Millennial native ad loaded.");
                         }
 
                         @Override
@@ -248,7 +250,7 @@ public class MillennialNative extends CustomEventNative {
         @Override
         public void onLoadFailed(NativeAd nativeAd, NativeAd.NativeErrorStatus nativeErrorStatus) {
             final NativeErrorCode error;
-            switch ( nativeErrorStatus.getErrorCode() ) {
+            switch (nativeErrorStatus.getErrorCode()) {
                 case NativeAd.NativeErrorStatus.LOAD_TIMED_OUT:
                     error = NativeErrorCode.NETWORK_TIMEOUT;
                     break;
@@ -278,24 +280,46 @@ public class MillennialNative extends CustomEventNative {
                     mListener.onNativeAdFailed(error);
                 }
             });
-            Log.i(LOGCAT_TAG, "Millennial native ad failed: " + nativeErrorStatus.getDescription() );
+            Log.i(TAG, "Millennial native ad failed: " + nativeErrorStatus.getDescription());
         }
 
         @Override
         public void onClicked(NativeAd nativeAd, NativeAd.ComponentName componentName, int i) {
-            Log.i(LOGCAT_TAG, "Millennial native SDK's click tracker fired.");
+            Log.d(TAG, "Millennial native ad click tracker fired.");
         }
 
         @Override
         public void onAdLeftApplication(NativeAd nativeAd) {
-            Log.i(LOGCAT_TAG, "Millennial native SDK has left the application.");
+            Log.d(TAG, "Millennial native ad has left the application.");
 
         }
 
         @Override
         public void onExpired(NativeAd nativeAd) {
-            Log.i(LOGCAT_TAG, "Millennial native ad has expired!");
+            Log.d(TAG, "Millennial native ad has expired!");
         }
 
+    }
+
+    private boolean initializeSDK(Context context) {
+        try {
+            if (android.os.Build.VERSION.SDK_INT >= Build.VERSION_CODES.JELLY_BEAN) {
+                if (!MMSDK.isInitialized()) {
+                    try {
+                        MMSDK.initialize((Activity) context, ActivityListenerManager.LifecycleState.RESUMED);
+                    } catch (Exception e) {
+                        Log.e(TAG, "Error initializing MMSDK", e);
+                        return false;
+                    }
+                }
+            } else {
+                Log.e(TAG, "MMSDK minimum supported API is 16");
+                return false;
+            }
+            return true;
+        } catch (Exception e) {
+            Log.e(TAG, "Error initializing MMSDK", e);
+            return false;
+        }
     }
 }