+++ mopub-sdk/mopub-sdk-base/src/main/java/com/mopub/common/util/ReflectionTarget.java
@@ -1,0 +1,7 @@
+package com.mopub.common.util;
+
+/**
+ * Methods that are accessed via reflection should be annotated with this so proguard does not
+ * obfuscate them.
+ */
+public @interface ReflectionTarget { }