+++ mopub-sdk/mopub-sdk-native-video/src/main/java/com/mopub/nativeads/NativeVideoViewController.java
@@ -260,7 +260,6 @@ public class NativeVideoViewController extends BaseVideoViewController implement
         applyState(newState);
     }
 
-
     @VisibleForTesting
     void applyState(@NonNull final VideoState videoState) {
         applyState(videoState, false);