+++ mopub-sdk/mopub-sdk-rewardedvideo/src/main/java/com/mopub/mobileads/MoPubRewardedVideos.java
@@ -6,6 +6,7 @@ import android.support.annotation.Nullable;
 
 import com.mopub.common.MediationSettings;
 import com.mopub.common.Preconditions;
+import com.mopub.common.util.ReflectionTarget;
 
 /**
  * MoPubRewardedVideos is a utility class that holds controller methods for other MoPub rewarded
@@ -13,6 +14,7 @@ import com.mopub.common.Preconditions;
  */
 public class MoPubRewardedVideos {
 
+    @ReflectionTarget
     public static void initializeRewardedVideo(@NonNull Activity activity,
             MediationSettings... mediationSettings) {
         Preconditions.checkNotNull(activity);
@@ -20,10 +22,12 @@ public class MoPubRewardedVideos {
         MoPubRewardedVideoManager.init(activity, mediationSettings);
     }
 
+    @ReflectionTarget
     public static void setRewardedVideoListener(@Nullable MoPubRewardedVideoListener listener) {
         MoPubRewardedVideoManager.setVideoListener(listener);
     }
 
+    @ReflectionTarget
     public static void loadRewardedVideo(@NonNull String adUnitId,
             @Nullable MediationSettings... mediationSettings) {
         Preconditions.checkNotNull(adUnitId);
@@ -31,6 +35,7 @@ public class MoPubRewardedVideos {
         MoPubRewardedVideoManager.loadVideo(adUnitId, null, mediationSettings);
     }
 
+    @ReflectionTarget
     public static void loadRewardedVideo(@NonNull String adUnitId,
             @Nullable MoPubRewardedVideoManager.RequestParameters requestParameters,
             @Nullable MediationSettings... mediationSettings) {
@@ -39,12 +44,14 @@ public class MoPubRewardedVideos {
         MoPubRewardedVideoManager.loadVideo(adUnitId, requestParameters, mediationSettings);
     }
 
+    @ReflectionTarget
     public static boolean hasRewardedVideo(@NonNull String adUnitId) {
         Preconditions.checkNotNull(adUnitId);
 
         return MoPubRewardedVideoManager.hasVideo(adUnitId);
     }
 
+    @ReflectionTarget
     public static void showRewardedVideo(@NonNull String adUnitId) {
         Preconditions.checkNotNull(adUnitId);